Replaced 'unsigned long long' with 'uint32_t' and 'uint64' with 'uint64_t' to avoid problems on 64-bit machines

+++ font.h
@@ -4,7 +4,7 @@
  * See the main source file 'vdr.c' for copyright information and
  * how to reach the author.
  *
- * $Id: font.h 1.12 2006/02/05 13:46:36 kls Exp $
+ * $Id: font.h 1.13 2006/12/02 11:11:48 kls Exp $
  */
 
 #ifndef __FONT_H
@@ -32,7 +32,7 @@ enum eDvbCode {
 class cFont {
 public:
   enum { NUMCHARS = 256 };
-  typedef unsigned long tPixelData;
+  typedef uint32_t tPixelData;
   struct tCharData {
     tPixelData width, height;
     tPixelData lines[1];
